Create ReadWriteKeywords
