+++ Mainframe3270/__init__.py
@@ -9,7 +9,7 @@ from robot.utils import ConnectionCache
 from robotlibcore import DynamicCore
 
 from Mainframe3270.keywords import (AssertionKeywords, ConnectionKeywords,
-                                    WaitAndTimeoutKeywords)
+                                    ReadWriteKeywords, WaitAndTimeoutKeywords)
 from Mainframe3270.py3270 import Emulator
 from Mainframe3270.utils import convert_timeout
 from Mainframe3270.version import VERSION
@@ -118,6 +118,7 @@ class Mainframe3270(DynamicCore):
             AssertionKeywords(self),
             ConnectionKeywords(self),
             WaitAndTimeoutKeywords(self),
+            ReadWriteKeywords(self),
             X3270(self),
         ]
         DynamicCore.__init__(self, libraries)
