+++ Mainframe3270/py3270.py
@@ -401,7 +401,8 @@ class Emulator(object):
         Co-ordinates are 1 based, as listed in the status area of the
         terminal.
         """
-        if xpos is not None and ypos is not None:
+        if xpos and ypos:
+            self._check_limits(ypos, xpos)
             self.move_to(ypos, xpos)
 
         # escape double quotes in the data to send
@@ -437,6 +438,11 @@ class Emulator(object):
         Co-ordinates are 1 based, as listed in the status area of the
         terminal.
         """
+        self._check_limits(ypos, xpos)
+        if (xpos + length) > (80 + 1):
+            raise Exception(
+                "You have exceeded the x-axis limit of the mainframe screen"
+            )
         # the screen's co-ordinates are 1 based, but the command is 0 based
         xpos -= 1
         ypos -= 1
@@ -494,3 +500,13 @@ class Emulator(object):
 
     def save_screen(self, file_path):
         self.exec_command("PrintText(html,file,{0})".format(file_path).encode("utf-8"))
+
+    def _check_limits(self, ypos, xpos):
+        if ypos > 24:
+            raise Exception(
+                "You have exceeded the y-axis limit of the mainframe screen"
+            )
+        if xpos > 80:
+            raise Exception(
+                "You have exceeded the x-axis limit of the mainframe screen"
+            )
